feat :  Refactoring and Ticket Classes
